TIP-882: move ref data repo resolver
