+++ src/Akeneo/Pim/Enrichment/Bundle/Doctrine/ReferenceDataRepositoryResolver.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\ReferenceDataBundle\Doctrine;
+namespace Akeneo\Pim\Enrichment\Bundle\Doctrine;
 
 use Akeneo\Pim\Enrichment\Component\Product\Repository\ReferenceDataRepositoryResolverInterface;
 use Akeneo\Pim\Structure\Component\ReferenceData\ConfigurationRegistryInterface;
