+++ src/Pim/Bundle/ReferenceDataBundle/Resources/config/services.yml
@@ -1,7 +1,7 @@
 parameters:
     pim_reference_data.registry.class: Akeneo\Pim\Structure\Component\ReferenceData\ConfigurationRegistry
     pim_reference_data.method_guesser.class: Akeneo\Pim\Enrichment\Component\Product\ReferenceData\MethodNameGuesser
-    pim_reference_data.repository_resolver.class: Pim\Bundle\ReferenceDataBundle\Doctrine\ReferenceDataRepositoryResolver
+    pim_reference_data.repository_resolver.class: Akeneo\Pim\Enrichment\Bundle\Doctrine\ReferenceDataRepositoryResolver
     pim_reference_data.label_renderer.class: Akeneo\Pim\Enrichment\Component\Product\ReferenceData\LabelRenderer
 
 services:
